Refactor insight map route handler into a testable library function
- Extracted logic from `src/app/api/insight/map/route.ts` to `src/lib/insight-map.ts`.
- Created `generateInsightMap` pure function for testability.
- Added unit tests in `src/lib/insight-map.test.ts`.
- Simplified route handler to focus on data fetching and response delivery.
